+++ nilearn/_utils/data_gen.py
@@ -341,7 +341,9 @@ def generate_fake_fmri_data_and_design(shapes, rk=3, affine=np.eye(4)):
         data = np.random.randn(*shape)
         data[1:-1, 1:-1, 1:-1] += 100
         fmri_data.append(Nifti1Image(data, affine))
-        columns = np.random.choice(list(string.ascii_lowercase), size=rk)
+        columns = np.random.choice(list(string.ascii_lowercase),
+                                   size=rk,
+                                   replace=False)
         design_matrices.append(pd.DataFrame(np.random.randn(shape[3], rk),
                                             columns=columns))
     mask = Nifti1Image((np.random.rand(*shape[:3]) > .5).astype(np.int8),